refactor(app): move the changelog viewer state into ainb-app
ChangelogState, its markdown line types and the embedded CHANGELOG.md move
to ainb_app::components::changelog; the viewer's renderer stays in
ainb-core.
